enhance the subscription to channels method
previously, we had a pseudo-subscribe, that is information of the channel was provided in the message. but not an actual subscription to channel - that is, the agents are still receiving messages from all channels but `choose` not to handle them in the `on_received_message`. 

now clear separation of channels is implemented and receiving end will not receive unintended messages from channels they have not subscribed to.